Add files via upload
Software version control is the practice of tracking changes to a software project's code and data over time. This allows developers to revert to previous versions of the code if necessary, and to collaborate on the project more effectively. There are many different software version control systems (VCS) available, but Git is one of the most popular.
Git is a distributed VCS, which means that each developer has a complete copy of the code repository on their local machine. This makes it possible for developers to work on the project offline, and to merge their changes with the central repository without having to be connected to the internet.
Git is also a very fast VCS, and it is designed to be used for managing large code repositories. This makes it a popular choice for open-source projects and for large corporations.